Add context to allow popup toolbar items to autohide
