+++ ui/appui-react/src/appui-react/widgets/ViewToolWidgetComposer.tsx
@@ -10,7 +10,7 @@ import classnames from "classnames";
 import * as React from "react";
 import { ToolbarOrientation, ToolbarUsage } from "@itwin/appui-abstract";
 import { ToolbarComposer } from "../toolbar/ToolbarComposer";
-import { useUiVisibility } from "./BasicToolWidget";
+import { useUiVisibility } from "../hooks/useUiVisibility";
 import { NavigationWidgetComposer } from "./NavigationWidgetComposer";
 
 /**
