+++ ui/components-react/src/components-react/toolbar/PopupItem.scss
@@ -14,14 +14,14 @@
   $arrow-shadow: $icon-shadow;
   position: relative;
 
-  >* {
+  > * {
     &:not(.components-triangle) {
       height: 100%;
       width: 100%;
     }
   }
 
-  >.components-triangle {
+  > .components-triangle {
     border-radius: 0;
     position: absolute;
     right: $margin-from-triangle-to-stroke;
@@ -29,10 +29,14 @@
     filter: $arrow-shadow;
     pointer-events: none;
 
-    @include triangle-bottom-right($width: $triangle-width, $height: $triangle-height, $color: $buic-icon-color);
+    @include triangle-bottom-right(
+      $width: $triangle-width,
+      $height: $triangle-height,
+      $color: $buic-icon-color
+    );
   }
 
-  &:active>.components-triangle {
+  &:active > .components-triangle {
     filter: $no-shadow;
 
     @include triangle-bottom-right-color($pressed-icon-color);
@@ -40,7 +44,7 @@
 
   &.components-active {
     &:not(:active):not(.components-disabled) {
-      >.components-triangle {
+      > .components-triangle {
         filter: $no-shadow;
 
         @include triangle-bottom-right-color($active-icon-color);
@@ -49,17 +53,17 @@
   }
 
   &.components-disabled {
-    >.components-triangle {
+    > .components-triangle {
       @include triangle-bottom-right-color($disabled-icon-color);
     }
   }
 
   &.components-disabled-drag {
-    >.components-icon {
+    > .components-icon {
       color: $disabled-icon-color;
     }
 
-    >.components-triangle {
+    > .components-triangle {
       @include triangle-bottom-right-color($disabled-icon-color);
     }
   }
@@ -73,4 +77,11 @@
   border-width: 1px;
   border-style: solid;
   border-radius: 3px;
+
+  transition: visibility 0.5s ease, opacity 0.5s ease;
+
+  &.nz-hidden {
+    opacity: 0;
+    visibility: hidden;
+  }
 }
