+++ ui/components-react/src/components-react/toolbar/PopupItem.tsx
@@ -15,6 +15,21 @@ import { ToolbarButtonItemProps } from "./Item";
 import { useToolbarWithOverflowDirectionContext, useToolItemEntryContext } from "./ToolbarWithOverflow";
 import { toToolbarPopupRelativePosition } from "./PopupItemWithDrag";
 
+/**
+ * Context used by Toolbar items to know if popup panel should be hidden - via AutoHide.
+ * @public
+ */
+// eslint-disable-next-line @typescript-eslint/naming-convention
+export const ToolbarPopupAutoHideContext = React.createContext<boolean>(false);
+
+/**
+ * React hook used to retrieve the ToolbarPopupAutoHideContext.
+ *  @public
+ */
+export function useToolbarPopupAutoHideContext() {
+  return React.useContext(ToolbarPopupAutoHideContext);
+}
+
 /** @public */
 export interface ToolbarPopupContextProps {
   readonly closePanel: () => void;
@@ -144,8 +159,13 @@ interface PopupItemPopupProps {
 
 /** @internal */
 export function PopupItemPopup(props: PopupItemPopupProps) {
+  const isHidden = useToolbarPopupAutoHideContext();
+  const className = classnames(
+    "components-toolbar-popupItem_popupItemPopup",
+    isHidden && "nz-hidden");
+
   return <Popup
-    className="components-toolbar-popupItem_popupItemPopup"
+    className={className}
     offset={0}
     showShadow={false}
     {...props}
